chore: refactored setup-passkey
